# docs/openapi/ai-service.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ai-service",
    "version": "0.1.0",
    "description": "Spring ↔ ai-service 내부 계약. 요청 = KAN-11 analyze() 출력 + focus + goal_amount. 처리가 끝나면 항상 200 + status, 입력 불일치만 422. POST /calculate 는 Kan-9 §2 입력 → §5 출력(승준 엔진). 상세: docs/KAN-17-내부HTTP계약.md"
  },
  "paths": {
...
    "/calculate": {
      "post": {
        "summary": "Calculate",
        "description": "Kan-9 §2 입력 dict 그대로 → 승준 analyze() → §5 출력 그대로 (M/Q/H 전부, 한 응답).\n\n검증은 엔진이 한다(정적 오류 여러 개 → 데이터 의존 오류 첫 것). ai-service 는 덧붙이지 않는다.\n  200  §5 출력 (status OK | NO_PLAN_FUNDS)\n  422  {\"code\":\"VALIDATION_ERROR\",\"errors\":[{code,field,message}…]}  — Spring이 400으로 변환해 그대로 전달\n  500  {\"code\":\"CALCULATION_FAILED\"}                                  — Spring은 502 CALCULATION_FAILED\n  503  {\"code\":\"ENGINE_UNAVAILABLE\"}                                  — engine/ 미배치·스냅샷 로드 실패",
        "operationId": "calculate_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanInputs"
              },
              "examples": {
                "P0": {
                  "value": {
                    "goal": {
                      "amount": 50000000,
                      "horizon_months": 60
                    },
                    "funds": {
                      "initial": 10000000,
                      "monthly": 600000
                    },
                    "alloc": {
                      "initial": {
                        "invest": 70,
                        "safe": 30,
                        "other": 0
                      },
                      "monthly": {
                        "invest": 50,
                        "safe": 40,
                        "other": 10
                      }
                    },
                    "portfolio": {
                      "assets": [
                        {
                          "code": "KR_EQ",
                          "weight": 40
                        },
                        {
                          "code": "US_EQ",
                          "weight": 40
                        },
                        {
                          "code": "KR_BOND",
                          "weight": 20
                        }
                      ]
                    },
                    "rebalancing": {
                      "focus": "Q"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Kan-9 §5 출력 (M/Q/H 전부)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calculation"
                }
              }
            }
          },
          "422": {
            "description": "엔진 검증 실패 — {code: VALIDATION_ERROR, errors: [{code, field, message}]}. Spring은 400으로 변환"
          },
          "500": {
            "description": "CALCULATION_FAILED — Spring은 502"
          },
          "503": {
            "description": "ENGINE_UNAVAILABLE — engine/ 미배치"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Period": {
        "description": "리밸런싱 주기. KAN-9 표기 M / Q / H.",
        "enum": [
          "M",
          "Q",
          "H"
        ],
        "title": "Period",
        "type": "string"
      },
...
      "AssetUsed": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "display_name": {
            "default": "",
            "title": "Display Name",
            "type": "string"
          },
          "instrument": {
            "default": "",
            "title": "Instrument",
            "type": "string"
          },
          "tax_class": {
            "default": "",
            "title": "Tax Class",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "AssetUsed",
        "type": "object"
      },
...
      "Derived": {
        "properties": {
          "propensity_label": {
            "default": "",
            "title": "Propensity Label",
            "type": "string"
          },
          "invest_share_overall_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Invest Share Overall Pct"
          },
          "plan_excluded_amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Plan Excluded Amount"
          }
        },
        "title": "Derived",
        "type": "object"
      },
      "Gap": {
        "properties": {
          "fv_total": {
            "title": "Fv Total",
            "type": "integer"
          },
          "shortfall": {
            "title": "Shortfall",
            "type": "integer"
          },
          "extra_monthly_required": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Extra Monthly Required"
          },
          "months_extension": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Months Extension"
          },
          "months_extension_raw": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Months Extension Raw"
          },
          "extension_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Extension Status"
          },
          "extra_monthly_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Extra Monthly Ratio"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Status"
          },
          "basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Basis"
          },
          "delta_m_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Delta M Model"
          }
        },
        "required": [
          "fv_total",
          "shortfall"
        ],
        "title": "Gap",
        "type": "object"
      },
      "Meta": {
        "properties": {
          "assumptions_version": {
            "default": "",
            "title": "Assumptions Version",
            "type": "string"
          },
          "data_version": {
            "default": "",
            "title": "Data Version",
            "type": "string"
          },
          "data_hash": {
            "default": "",
            "title": "Data Hash",
            "type": "string"
          },
          "window": {
            "$ref": "#/components/schemas/Window"
          },
          "assets_used": {
            "items": {
              "$ref": "#/components/schemas/AssetUsed"
            },
            "title": "Assets Used",
            "type": "array"
          },
          "data_basis": {
            "default": "",
            "title": "Data Basis",
            "type": "string"
          },
          "generated_at": {
            "default": "",
            "title": "Generated At",
            "type": "string"
          },
          "safe_rate_annual_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Safe Rate Annual Pct"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "start_month": {
            "default": "",
            "title": "Start Month",
            "type": "string"
          },
          "target_month": {
            "default": "",
            "title": "Target Month",
            "type": "string"
          },
          "cashflow_source": {
            "default": "",
            "title": "Cashflow Source",
            "type": "string"
          },
          "series_used": {
            "items": {
              "type": "string"
            },
            "title": "Series Used",
            "type": "array"
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Options"
          }
        },
        "required": [
          "window"
        ],
        "title": "Meta",
        "type": "object"
      },
      "PeriodResult": {
        "properties": {
          "trajectory": {
            "items": {
              "$ref": "#/components/schemas/TrajectoryPoint"
            },
            "title": "Trajectory",
            "type": "array"
          },
          "cum_cost": {
            "title": "Cum Cost",
            "type": "integer"
          },
          "risk": {
            "$ref": "#/components/schemas/Risk"
          },
          "gap": {
            "$ref": "#/components/schemas/Gap"
          },
          "tax": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Tax"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "cum_cost",
          "risk",
          "gap"
        ],
        "title": "PeriodResult",
        "type": "object"
      },
      "Risk": {
        "properties": {
          "mdd_pct": {
            "title": "Mdd Pct",
            "type": "number"
          },
          "vol_annual_pct": {
            "title": "Vol Annual Pct",
            "type": "number"
          },
          "worst_month_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Worst Month Pct"
          },
          "max_drift_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Max Drift Pct"
          }
        },
        "required": [
          "mdd_pct",
          "vol_annual_pct"
        ],
        "title": "Risk",
        "type": "object"
      },
...
      "Tax": {
        "properties": {
          "realized_cum": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Realized Cum"
          },
          "fv_after_tax": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Fv After Tax"
          }
        },
        "title": "Tax",
        "type": "object"
      },
      "TrajectoryPoint": {
        "properties": {
          "month": {
            "title": "Month",
            "type": "integer"
          },
          "invest": {
            "default": 0,
            "title": "Invest",
            "type": "integer"
          },
          "safe": {
            "default": 0,
            "title": "Safe",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "month",
          "total"
        ],
        "title": "TrajectoryPoint",
        "type": "object"
      },
      "Window": {
        "properties": {
          "start": {
            "title": "Start",
            "type": "string"
          },
          "end": {
            "title": "End",
            "type": "string"
          },
          "months": {
            "title": "Months",
            "type": "integer"
          }
        },
        "required": [
          "start",
          "end",
          "months"
        ],
        "title": "Window",
        "type": "object"
      },
      "Alloc": {
        "additionalProperties": false,
        "properties": {
          "initial": {
            "$ref": "#/components/schemas/AllocSplit",
            "description": "합 100 아니면 ALLOC_SUM_INITIAL"
          },
          "monthly": {
            "$ref": "#/components/schemas/AllocSplit",
            "description": "합 100 아니면 ALLOC_SUM_MONTHLY"
          }
        },
        "required": [
          "initial",
          "monthly"
        ],
        "title": "Alloc",
        "type": "object"
      },
      "AllocSplit": {
        "additionalProperties": false,
        "description": "자금 배분율. 정수 퍼센트, 합 100. '투자 성향' 입력은 없고 이 배분율에서 파생된다 (Kan-9 확정 ⑪).",
        "properties": {
          "invest": {
            "maximum": 100,
            "minimum": 0,
            "title": "Invest",
            "type": "integer"
          },
          "safe": {
            "maximum": 100,
            "minimum": 0,
            "title": "Safe",
            "type": "integer"
          },
          "other": {
            "maximum": 100,
            "minimum": 0,
            "title": "Other",
            "type": "integer"
          }
        },
        "required": [
          "invest",
          "safe",
          "other"
        ],
        "title": "AllocSplit",
        "type": "object"
      },
      "AssetWeight": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Kan-9 카탈로그 코드 (KR_EQ · US_EQ · KR_BOND · US_EQ_KR). 밖이면 ASSET_NOT_IN_CATALOG",
            "title": "Code",
            "type": "string"
          },
          "weight": {
            "description": "정수 퍼센트. 0 < w ≤ 100 아니면 PORTFOLIO_WEIGHT_RANGE",
            "maximum": 100,
            "minimum": 1,
            "title": "Weight",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "weight"
        ],
        "title": "AssetWeight",
        "type": "object"
      },
      "Calculation": {
        "description": "ai-service POST /calculate 응답 = KAN-11 analyze() 출력 그대로 (Kan-9 §5). 백엔드는 가공·저장하지 않는다.\n\n`schema.SimulationInput` 과 같은 모양에서 `focus`·`goal_amount` 만 없다 — 그 둘은 Spring이\nPOST /rag/answer 를 부를 때 덧붙이는 값이라 공개 응답에는 나오지 않는다.",
        "properties": {
          "status": {
            "default": "OK",
            "title": "Status",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "derived": {
            "$ref": "#/components/schemas/Derived"
          },
          "per_period": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PeriodResult"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Period"
            },
            "title": "Per Period",
            "type": "object"
          }
        },
        "required": [
          "meta",
          "per_period"
        ],
        "title": "Calculation",
        "type": "object"
      },
      "Funds": {
        "additionalProperties": false,
        "properties": {
          "initial": {
            "description": "초기 투자금 KRW. 음수 → FUNDS_INITIAL_RANGE",
            "minimum": 0,
            "title": "Initial",
            "type": "integer"
          },
          "monthly": {
            "description": "월 납입액 KRW (상수). 음수 → FUNDS_MONTHLY_RANGE. initial·monthly 동시 0 → NO_FUNDS",
            "minimum": 0,
            "title": "Monthly",
            "type": "integer"
          }
        },
        "required": [
          "initial",
          "monthly"
        ],
        "title": "Funds",
        "type": "object"
      },
      "Goal": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "description": "목표 금액 KRW. 범위 밖 → GOAL_AMOUNT_RANGE",
            "maximum": 10000000000,
            "minimum": 1000000,
            "title": "Amount",
            "type": "integer"
          },
          "horizon_months": {
            "description": "목표 기간(개월). 범위 밖 → GOAL_HORIZON_RANGE",
            "maximum": 120,
            "minimum": 12,
            "title": "Horizon Months",
            "type": "integer"
          }
        },
        "required": [
          "amount",
          "horizon_months"
        ],
        "title": "Goal",
        "type": "object"
      },
      "PlanInputs": {
        "additionalProperties": false,
        "description": "POST /plans 요청 본문 = Kan-9 §2 입력 dict 그대로. Spring은 변환 없이 ai-service POST /calculate 에 넘긴다.",
        "properties": {
          "goal": {
            "$ref": "#/components/schemas/Goal"
          },
          "funds": {
            "$ref": "#/components/schemas/Funds"
          },
          "alloc": {
            "$ref": "#/components/schemas/Alloc"
          },
          "portfolio": {
            "$ref": "#/components/schemas/Portfolio"
          },
          "rebalancing": {
            "$ref": "#/components/schemas/Rebalancing"
          }
        },
        "required": [
          "goal",
          "funds",
          "alloc",
          "portfolio",
          "rebalancing"
        ],
        "title": "PlanInputs",
        "type": "object"
      },
      "Portfolio": {
        "additionalProperties": false,
        "properties": {
          "assets": {
            "description": "1~3개, weight 합 100 (WEIGHTS_SUM), code 중복 불가 (PORTFOLIO_ASSET_DUP)",
            "items": {
              "$ref": "#/components/schemas/AssetWeight"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Assets",
            "type": "array"
          }
        },
        "required": [
          "assets"
        ],
        "title": "Portfolio",
        "type": "object"
      },
      "Rebalancing": {
        "additionalProperties": false,
        "properties": {
          "focus": {
            "$ref": "#/components/schemas/Period",
            "description": "강조 주기 M/Q/H. 필수 — FOCUS_INVALID. 계산은 세 주기 전부, 화면·AI에서 강조"
          }
        },
        "required": [
          "focus"
        ],
        "title": "Rebalancing",
        "type": "object"
      }
    }
  }
}